requirements.txt
streamlit==1.45.1
opencv-python==4.11.0.86
pillow==11.2.1
numpy==1.26.4 